+++ packages/api/src/decorator/private/ActivityDecorator.tsx
@@ -1,4 +1,4 @@
-import { WebChatActivity } from 'botframework-webchat-core';
+import { getActivityLivestreamingMetadata, type WebChatActivity } from 'botframework-webchat-core';
 import React, { Fragment, memo, useMemo, type ReactNode } from 'react';
 import { ActivityDecoratorRequest } from '..';
 import { ActivityBorderDecoratorMiddlewareProxy } from './ActivityBorderDecoratorMiddleware';
@@ -8,18 +8,23 @@ const ActivityDecoratorFallback = memo(({ children }) => <Fragment>{children}</F
 ActivityDecoratorFallback.displayName = 'ActivityDecoratorFallback';
 
 const supportedActivityRoles: ActivityDecoratorRequest['from'][] = ['bot', 'channel', 'user', undefined];
-const supportedActivityStates: ActivityDecoratorRequest['state'][] = ['informative', 'completion', undefined];
 
-function ActivityDecorator({ children, activity }: Readonly<{ activity?: WebChatActivity; children?: ReactNode }>) {
-  const request = useMemo<ActivityDecoratorRequest>(
-    () => ({
-      from: supportedActivityRoles.includes(activity?.from?.role) ? activity?.from?.role : undefined,
-      state: supportedActivityStates.includes(activity?.channelData?.streamType)
-        ? activity?.channelData?.streamType
-        : undefined
-    }),
-    [activity]
-  );
+function ActivityDecorator({ activity, children }: Readonly<{ activity?: WebChatActivity; children?: ReactNode }>) {
+  const request = useMemo<ActivityDecoratorRequest>(() => {
+    const { type } = getActivityLivestreamingMetadata(activity) || {};
+
+    return {
+      livestreamingState:
+        type === 'final activity'
+          ? 'completing'
+          : type === 'informative message'
+            ? 'preparing'
+            : type === 'interim activity'
+              ? 'ongoing'
+              : undefined,
+      from: supportedActivityRoles.includes(activity?.from?.role) ? activity?.from?.role : undefined
+    };
+  }, [activity]);
 
   return (
     <ActivityBorderDecoratorMiddlewareProxy fallbackComponent={ActivityDecoratorFallback} request={request}>